// packages/headless/src/builders/brick-builder/types.ts
export type FieldTypes =
	| "tab"
	| "text"
	| "wysiwyg"
	| "media"
	| "repeater"
	| "number"
	| "checkbox"
	| "select"
	| "textarea"
	| "json"
	| "colour"
	| "datetime"
	| "pagelink"
	| "link";